Broken up and fixed ServerRequest tests
