+++ tests/ServerRequest/BodyTest.php
@@ -1,0 +1,47 @@
+<?php
+
+namespace Jasny\HttpMessage\ServerRequest;
+
+use PHPUnit_Framework_TestCase;
+use Jasny\TestHelper;
+
+use Jasny\HttpMessage\ServerRequest;
+use Jasny\HttpMessage\Stream;
+
+/**
+ * @covers Jasny\HttpMessage\Message\Body
+ * @covers Jasny\HttpMessage\ServerRequest\Body
+ */
+class BodyTest extends PHPUnit_Framework_TestCase
+{
+    use TestHelper;
+    
+    /**
+     * @var ServerRequest
+     */
+    protected $baseRequest;
+    
+    public function setUp()
+    {
+        $this->baseRequest = new ServerRequest();
+    }
+
+    public function testGetBodyDefault()
+    {
+        $body = $this->baseRequest->getBody();
+        
+        $this->assertInstanceOf(Stream::class, $body);
+        $this->assertEquals('data://text/plain,', $body->getMetadata('uri'));
+    }
+
+    public function testWithBody()
+    {
+        $stream = $this->createMock(Stream::class);
+        $request = $this->baseRequest->withBody($stream);
+        
+        $this->assertInstanceof(ServerRequest::class, $request);
+        $this->assertNotSame($this->baseRequest, $request);
+        
+        $this->assertSame($stream, $request->getBody());
+    }
+}
