+++ tests/ServerRequest/MethodTest.php
@@ -1,0 +1,66 @@
+<?php
+
+namespace Jasny\HttpMessage\ServerRequest;
+
+use PHPUnit_Framework_TestCase;
+use Jasny\TestHelper;
+
+use Jasny\HttpMessage\ServerRequest;
+
+/**
+ * @covers Jasny\HttpMessage\ServerRequest\Method
+ */
+class MethodTest extends PHPUnit_Framework_TestCase
+{
+    use TestHelper;
+    
+    /**
+     * @var ServerRequest
+     */
+    protected $baseRequest;
+    
+    public function setUp()
+    {
+        $this->baseRequest = new ServerRequest();
+    }
+
+
+    public function testGetMethodDefault()
+    {
+        $this->assertSame('', $this->baseRequest->getMethod());
+    }
+
+    public function testDetermineMethod()
+    {
+        $request = $this->baseRequest->withServerParams(['REQUEST_METHOD' => 'post']);
+        $this->assertEquals('POST', $request->getMethod());
+    }
+
+    public function testWithMethod()
+    {
+        $request = $this->baseRequest->withMethod('Get'); // Case should not be changed
+        
+        $this->assertInstanceof(ServerRequest::class, $request);
+        $this->assertNotSame($this->baseRequest, $request);
+        
+        $this->assertEquals('Get', $request->getMethod());
+    }
+
+    /**
+     * @expectedException InvalidArgumentException
+     * @expectedExceptionMessage Invalid method 'foo bar': Method may only contain letters and dashes
+     */
+    public function testWithMethodInvalid()
+    {
+        $this->baseRequest->withMethod("foo bar");
+    }
+
+    /**
+     * @expectedException InvalidArgumentException
+     * @expectedExceptionMessage Method should be a string, not a stdClass object
+     */
+    public function testWithMethodInvalidArgument()
+    {
+        $this->baseRequest->withMethod((object)['foo' => 1, 'bar' => 2]);
+    }
+}
