+++ tests/ServerRequest/ProtocolVersionTest.php
@@ -1,0 +1,78 @@
+<?php
+
+namespace Jasny\HttpMessage\ServerRequest;
+
+use PHPUnit_Framework_TestCase;
+use Jasny\TestHelper;
+
+use Jasny\HttpMessage\ServerRequest;
+
+/**
+ * @covers Jasny\HttpMessage\Message\ProtocolVersion
+ * @covers Jasny\HttpMessage\ServerRequest\ProtocolVersion
+ */
+class ServerParamsTest extends PHPUnit_Framework_TestCase
+{
+    use TestHelper;
+    
+    /**
+     * @var ServerRequest
+     */
+    protected $baseRequest;
+    
+    public function setUp()
+    {
+        $this->baseRequest = new ServerRequest();
+    }
+    
+
+    public function testDefaultProtocolVersion()
+    {
+        $this->assertEquals('1.1', $this->baseRequest->getProtocolVersion());
+    }
+
+    public function testDetermineProtocolVersion()
+    {
+        $request = $this->baseRequest->withServerParams(['SERVER_PROTOCOL' => 'HTTP/1.0']);
+        
+        $this->assertEquals('1.0', $request->getProtocolVersion());
+    }
+
+    public function testWithProtocolVersion()
+    {
+        $request = $this->baseRequest->withProtocolVersion('1.1');
+        
+        $this->assertInstanceof(ServerRequest::class, $request);
+        $this->assertNotSame($this->baseRequest, $request);
+        
+        $this->assertEquals('1.1', $request->getProtocolVersion());
+    }
+
+    public function testWithProtocolVersionFloat()
+    {
+        $request = $this->baseRequest->withProtocolVersion(2.0);
+        
+        $this->assertInstanceof(ServerRequest::class, $request);
+        $this->assertNotSame($this->baseRequest, $request);
+        
+        $this->assertEquals('2.0', $request->getProtocolVersion());
+    }
+
+    /**
+     * @expectedException InvalidArgumentException
+     * @expectedExceptionMessage Invalid HTTP protocol version '0.2'
+     */
+    public function testWithInvalidProtocolVersion()
+    {
+        $this->baseRequest->withProtocolVersion('0.2');
+    }
+
+    /**
+     * @expectedException InvalidArgumentException
+     * @expectedExceptionMessage HTTP version must be a string or float
+     */
+    public function testWithInvalidTypeProtocolVersion()
+    {
+        $this->baseRequest->withProtocolVersion([]);
+    }
+}
