+++ tests/ServerRequest/QueryParamsTest.php
@@ -1,0 +1,54 @@
+<?php
+
+namespace Jasny\HttpMessage\ServerRequest;
+
+use PHPUnit_Framework_TestCase;
+use Jasny\TestHelper;
+
+use Jasny\HttpMessage\ServerRequest;
+
+/**
+ * @covers Jasny\HttpMessage\ServerRequest\QueryParams
+ */
+class QueryParamsTest extends PHPUnit_Framework_TestCase
+{
+    use TestHelper;
+    
+    /**
+     * @var ServerRequest
+     */
+    protected $baseRequest;
+    
+    public function setUp()
+    {
+        $this->baseRequest = new ServerRequest();
+    }
+    
+    
+    public function testGetQueryParamsDefault()
+    {
+        $this->assertSame([], $this->baseRequest->getQueryParams());
+    }
+
+    public function testWithQueryParams()
+    {
+        $request = $this->baseRequest->withQueryParams(['foo' => 'bar', 'color' => 'red']);
+        
+        $this->assertInstanceof(ServerRequest::class, $request);
+        $this->assertNotSame($this->baseRequest, $request);
+        
+        $this->assertSame(['foo' => 'bar', 'color' => 'red'], $request->getQueryParams());
+    }
+
+    public function testWithQueryParamsTurnStale()
+    {
+        $refl = new \ReflectionProperty(ServerRequest::class, 'isStale');
+        $refl->setAccessible(true);
+        $refl->setValue($this->baseRequest, false);
+        
+        $request = $this->baseRequest->withQueryParams([]);
+        
+        $this->assertTrue($this->baseRequest->isStale());
+        $this->assertFalse($request->isStale());
+    }
+}
