+++ tests/ServerRequest/RequestTargetTest.php
@@ -1,0 +1,63 @@
+<?php
+
+namespace Jasny\HttpMessage\ServerRequest;
+
+use PHPUnit_Framework_TestCase;
+use Jasny\TestHelper;
+
+use Jasny\HttpMessage\ServerRequest;
+
+/**
+ * @covers Jasny\HttpMessage\ServerRequest\RequestTarget
+ */
+class RequestTargetTest extends PHPUnit_Framework_TestCase
+{
+    use TestHelper;
+    
+    /**
+     * @var ServerRequest
+     */
+    protected $baseRequest;
+    
+    public function setUp()
+    {
+        $this->baseRequest = new ServerRequest();
+    }
+    
+
+    public function testGetRequestTargetDefault()
+    {
+        $this->assertEquals('/', $this->baseRequest->getRequestTarget());
+    }
+
+    public function testDetermineRequestTarget()
+    {
+        $request = $this->baseRequest->withServerParams(['REQUEST_URI' => '/foo?bar=1']);
+        $this->assertEquals('/foo?bar=1', $request->getRequestTarget());
+    }
+
+    public function testDetermineRequestTargetOrigin()
+    {
+        $request = $this->baseRequest->withServerParams(['REQUEST_METHOD' => 'OPTIONS']);
+        $this->assertEquals('*', $request->getRequestTarget());
+    }
+
+    public function testWithRequestTarget()
+    {
+        $request = $this->baseRequest->withRequestTarget('/foo?bar=99');
+        
+        $this->assertInstanceof(ServerRequest::class, $request);
+        $this->assertNotSame($this->baseRequest, $request);
+        
+        $this->assertEquals('/foo?bar=99', $request->getRequestTarget());
+    }
+
+    /**
+     * @expectedException InvalidArgumentException
+     * @expectedExceptionMessage Request target should be a string, not a stdClass object
+     */
+    public function testWithRequestTargetWithInvalidArgument()
+    {
+        $this->baseRequest->withRequestTarget((object)['foo' => 1, 'bar' => 2, 'zoo' => 3]);
+    }
+}
